สอบโปรเจค1

+++ resources/views/backend/position/index.blade.php
@@ -23,7 +23,7 @@
                                 <a href="{{ url('/bposition/'.$position->id.'/edit') }}" class="btn btn-success">แก้ไข</a>
                             </td>
                             <td>
-                                <?= Form::open(array('url' => 'bposition/' . $position->id, 'method' => 'delete')) ?>
+                            <a onclick="return confirm('คุณต้องการที่จะลบใช่หรือไม่?');"><?= Form::open(array('url' => 'bposition/' . $position->id, 'method' => 'delete')) ?>
                                 <button type="submit" class="btn btn-danger">ลบ</button>
                                 {!! Form::close() !!}
                             </td>
