+++ resources/views/backend/studentyear/index.blade.php
@@ -5,14 +5,14 @@
     <div class="row justify-content-center">
         <div class="col-md-10 col-md-offset-1">
             <br><br>
-            <?= link_to('bstudentyear/create', $title = 'เพิ่มข้อมูลรุ่น', ['class' => 'btn btn-success btn-sm'], $secure = null); ?>
+            <?= link_to('bstudentyear/create', $title = 'เพิ่มข้อมูลปีที่เข้าศึกษา', ['class' => 'btn btn-success btn-sm'], $secure = null); ?>
             <hr>
             <div class="card">
 
                 <div class="card-body">
                     <table class="table table-striped">
                         <tr>
-                            <th>รุ่น</th>
+                            <th>ปีที่เข้าศึกษา</th>
                             <th>แก้ไข</th>
                             <th>ลบ</th>
                         </tr>
@@ -23,7 +23,7 @@
                                 <a href="{{ url('/bstudentyear/'.$studentyear->id.'/edit') }}" class="btn btn-success">แก้ไข</a>
                             </td>
                             <td>
-                                <?= Form::open(array('url' => 'bstudentyear/' . $studentyear->id, 'method' => 'delete')) ?>
+                                <a onclick="return confirm('คุณต้องการที่จะลบใช่หรือไม่?');"><?= Form::open(array('url' => 'bstudentyear/' . $studentyear->id, 'method' => 'delete')) ?>
                                 <button type="submit" class="btn btn-danger">ลบ</button>
                                 {!! Form::close() !!}
                             </td>
